move core into sourmash
