+++ Cargo.toml
@@ -1,4 +1,18 @@
 [workspace]
-members = ["cli", "server", "server/frontend", "core"]
+members = ["cli", "server", "server/frontend", "p2p"]
 
 default-members = ["cli"]
+
+[profile.release]
+debug = true
+#[profile.release]
+# less code to include into binary
+#panic = 'abort'
+# optimization over all codebase ( better optimization, slower build )
+#codegen-units = 1
+# optimization for size ( more aggressive )
+#opt-level = 'z'
+# optimization for size
+# opt-level = 's'
+# link time optimization using using whole-program analysis
+#lto = true
