+++ cli/Cargo.toml
@@ -10,8 +10,7 @@ edition = "2018"
 structopt = "0.3.15"
 env_logger = "0.7.1"
 log = "0.4.8"
-greyhound-core = { path = "../core" }
-sourmash = "0.10.0"
+sourmash = { git = "https://github.com/dib-lab/sourmash.git", branch = "greyhound", features = ["experimental"]}
 rayon = "1.0"
 serde_json = "1.0.56"
 niffler = "2.2.0"
