--- core/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "greyhound-core"
-version = "0.1.0"
-authors = ["Luiz Irber <luiz.irber@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-nohash-hasher = "0.2.0"
-rayon = "1.3.1"
-serde = { version = "1.0.105", features = ["derive"] }
-sourmash = "0.10.0"
-serde_json = "1.0.56"
-niffler = "2.2.0"
-log = "0.4.8"
-counter = "0.5.2"
