+++ server/frontend/Cargo.toml
@@ -21,7 +21,8 @@ wasm-logger = "0.2"
 serde = { version = "1.0", features = ["derive"] }
 niffler = { version = "2.2.0", default-features = false, features = ["gz"] }
 needletail = { version = "0.4.0", default-features = false }
-sourmash = "0.10.0"
+#sourmash = { git = "https://github.com/dib-lab/sourmash.git", branch = "greyhound", features = ["experimental"]}
+sourmash = { path = "../../../sourmash/src/core", features = ["experimental"]}
 wee_alloc = "0.4.5"
 web-sys = "0.3.45"
 js-sys = "0.3"
