+++ server/frontend/src/lib.rs
@@ -4,55 +4,31 @@ pub mod native_worker;
 
 use anyhow::Error;
 use log::info;
-use needletail::{parse_fastx_reader, Sequence};
-use serde::{Deserialize, Serialize};
 use web_sys::DragEvent;
-use yew::format::{Json, Nothing};
+use yew::format::Json;
 use yew::services::fetch::{FetchService, FetchTask, Request, Response};
-use yew::services::reader::{File, FileChunk, FileData, ReaderService, ReaderTask};
+use yew::services::reader::{File, FileData, ReaderService, ReaderTask};
 use yew::worker::{Bridge, Bridged};
 use yew::{html, Callback, ChangeData, Component, ComponentLink, Html, ShouldRender};
 
-use sourmash::cmd::ComputeParameters;
+use sourmash::index::greyhound::GatherResult;
 use sourmash::signature::Signature;
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GatherResult {
-    intersect_bp: usize,
-    f_orig_query: f64,
-    f_match: f64,
-    f_unique_to_query: f64,
-    f_unique_weighted: f64,
-    average_abund: usize,
-    median_abund: usize,
-    std_abund: usize,
-    filename: String,
-    name: String,
-    md5: String,
-    match_: String,
-    f_match_orig: f64,
-    unique_intersect_bp: usize,
-    gather_result_rank: usize,
-    remaining_bp: usize,
-}
 
 pub struct Model {
     link: ComponentLink<Self>,
     job: Box<dyn Bridge<native_worker::Worker>>,
-    reader: ReaderService,
-    tasks: Vec<ReaderTask>,
-    files: Vec<String>,
     ft: Option<FetchTask>,
     sig: Option<Signature>,
+    reader: ReaderService,
+    tasks: Vec<ReaderTask>,
 }
 
 pub enum Msg {
-    SendToWorker,
-    DataReceived,
-    Drop(DragEvent),
-    Loaded(FileData),
-    Chunk(Option<FileChunk>),
+    SendToWorker(Vec<u8>),
     Files(Vec<File>),
+    Loaded(FileData),
+    DataReceived(Vec<u8>),
+    Drop(DragEvent),
     FetchData(Vec<u8>),
     FetchReady(Result<Vec<GatherResult>, Error>),
     Ignore,
@@ -63,29 +39,38 @@ impl Component for Model {
     type Properties = ();
 
     fn create(_: Self::Properties, link: ComponentLink<Self>) -> Self {
-        let callback = link.callback(|_| Msg::DataReceived);
+        let callback = link.callback(|m: native_worker::Response| match m {
+            native_worker::Response::Signature(sig) => Msg::DataReceived(sig),
+        });
         let job = native_worker::Worker::bridge(callback);
 
         Model {
             link,
             job,
-            reader: ReaderService::new(),
-            tasks: vec![],
-            files: vec![],
             ft: None,
             sig: None,
+            reader: ReaderService::new(),
+            tasks: vec![],
         }
     }
 
     fn update(&mut self, msg: Self::Message) -> ShouldRender {
         match msg {
-            Msg::SendToWorker => {
-                self.job.send(native_worker::Request::GetDataFromServer);
+            Msg::SendToWorker(raw_data) => {
+                self.job.send(native_worker::Request::ProcessFile(raw_data));
             }
-            Msg::DataReceived => {
-                info!("DataReceived");
+            Msg::DataReceived(sig) => {
+                self.sig = Some(Signature::from_reader(&sig[..]).unwrap().swap_remove(0));
+                self.link.send_message(Msg::FetchData(sig));
             }
             Msg::Drop(_) => unimplemented!(),
+            Msg::Loaded(file) => {
+                let mut buf = vec![];
+                let (mut reader, _) = niffler::get_reader(Box::new(&file.content[..])).unwrap();
+                reader.read_to_end(&mut buf).unwrap();
+
+                self.link.send_message(Msg::SendToWorker(buf));
+            }
             Msg::FetchData(json) => {
                 let callback = self.link.callback(
                     move |response: Response<Json<Result<Vec<GatherResult>, Error>>>| {
@@ -106,43 +91,11 @@ impl Component for Model {
                 // result is Vec<GatherResult>
                 //todo!("populate the table")
             }
-            Msg::Loaded(file) => {
-                let params = ComputeParameters::builder()
-                    .ksizes(vec![21])
-                    .num_hashes(0)
-                    .scaled(2000)
-                    .build();
-                let mut sig = Signature::from_params(&params);
-
-                let mut buf = vec![];
-                let (mut reader, _) = niffler::get_reader(Box::new(&file.content[..])).unwrap();
-                reader.read_to_end(&mut buf).unwrap();
-                let mut parser = parse_fastx_reader(&buf[..]).unwrap();
-                while let Some(record) = parser.next() {
-                    let record = record.unwrap();
-                    let norm_seq = record.normalize(true);
-                    sig.add_sequence(&norm_seq, true).unwrap();
-                }
-                let json = serde_json::to_vec(&[&sig]).unwrap();
-                self.sig = Some(sig);
-                // info!("{:?}", &json);
-                self.link.send_message(Msg::FetchData(json));
-            }
-            Msg::Chunk(Some(chunk)) => {
-                let info = format!("chunk: {:?}", chunk);
-                self.files.push(info);
-            }
             Msg::Files(files) => {
-                let chunks = false;
                 for file in files.into_iter() {
                     let task = {
-                        if chunks {
-                            let callback = self.link.callback(Msg::Chunk);
-                            self.reader.read_file_by_chunks(file, callback, 10).unwrap()
-                        } else {
-                            let callback = self.link.callback(Msg::Loaded);
-                            self.reader.read_file(file, callback).unwrap()
-                        }
+                        let callback = self.link.callback(Msg::Loaded);
+                        self.reader.read_file(file, callback).unwrap()
                     };
                     self.tasks.push(task);
                 }
